refactor(module:comment): refactor and aync Ant Design 4
ref #4634
