+++ components/comment/comment.component.ts
@@ -1,0 +1,65 @@
+/**
+ * @license
+ * Copyright Alibaba.com All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://github.com/NG-ZORRO/ng-zorro-antd/blob/master/LICENSE
+ */
+
+import { ChangeDetectionStrategy, Component, ContentChildren, Input, QueryList, TemplateRef, ViewEncapsulation } from '@angular/core';
+
+import { NzCommentActionComponent as CommentAction } from './comment-cells';
+
+@Component({
+  selector: 'nz-comment',
+  exportAs: 'nzComment',
+  template: `
+    <div class="ant-comment-inner">
+      <div class="ant-comment-avatar">
+        <ng-content select="nz-avatar[nz-comment-avatar]"></ng-content>
+      </div>
+      <div class="ant-comment-content">
+        <div class="ant-comment-content-author">
+          <span *ngIf="nzAuthor" class="ant-comment-content-author-name">
+            <ng-container *nzStringTemplateOutlet="nzAuthor">{{ nzAuthor }}</ng-container>
+          </span>
+          <span *ngIf="nzDatetime" class="ant-comment-content-author-time">
+            <ng-container *nzStringTemplateOutlet="nzDatetime">{{ nzDatetime }}</ng-container>
+          </span>
+        </div>
+        <ng-content select="nz-comment-content"></ng-content>
+        <ul class="ant-comment-actions" *ngIf="actions?.length">
+          <li *ngFor="let action of actions">
+            <span><ng-template [nzCommentActionHost]="action.content"></ng-template></span>
+          </li>
+        </ul>
+      </div>
+    </div>
+    <div class="ant-comment-nested">
+      <ng-content></ng-content>
+    </div>
+  `,
+  encapsulation: ViewEncapsulation.None,
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  host: {
+    class: 'ant-comment'
+  },
+  styles: [
+    `
+      nz-comment {
+        display: block;
+      }
+
+      nz-comment-content {
+        display: block;
+      }
+    `
+  ]
+})
+export class NzCommentComponent {
+  @Input() nzAuthor: string | TemplateRef<void>;
+  @Input() nzDatetime: string | TemplateRef<void>;
+
+  @ContentChildren(CommentAction) actions: QueryList<CommentAction>;
+  constructor() {}
+}
